+++ src/app/[locale]/(logged-out)/about/cohort-discovery/page.tsx
@@ -1,3 +1,4 @@
+import { notFound } from "next/navigation";
 import CohortDiscoveryInfo from "@/components/CohortDiscoveryInfo";
 import { getNewCohortDiscovery } from "@/utils/cms";
 import metaData from "@/utils/metadata";
@@ -9,6 +10,10 @@ export const metadata = metaData({
 
 export default async function CohortDiscoveryPage() {
     const cohortDiscovery = await getNewCohortDiscovery();
+
+    if (!cohortDiscovery) {
+        return notFound();
+    }
 
     return (
         <CohortDiscoveryInfo