+++ src/components/IndicateNhseSdeAccessButton/IndicateNhseSdeAccessButton.tsx
@@ -1,12 +1,14 @@
 "use client";
 
 import { ReactElement, useState } from "react";
-import { CircularProgress, SxProps, Tooltip } from "@mui/material";
+import { CircularProgress, Stack, SxProps, Tooltip } from "@mui/material";
 import { useTranslations } from "next-intl";
+import Box from "@/components/Box";
+import Typography from "@/components/Typography";
 import useAuth from "@/hooks/useAuth";
 import { useCohortStatus } from "@/hooks/useCohortStatus";
+import useModal from "@/hooks/useModal";
 import usePost from "@/hooks/usePost";
-import notificationService from "@/services/notification";
 import apis from "@/config/apis";
 import { revalidateCacheAction } from "@/app/actions/revalidateCacheAction";
 import { useFeatures } from "@/providers/FeatureProvider";
@@ -16,14 +18,25 @@ import ConditionalWrapper from "../ConditionalWrapper";
 const tooltipWrapper = (tooltip: string) => (children: ReactElement) =>
     (
         <Tooltip title={tooltip} describeChild placement="bottom">
-            {children}
+            <Box sx={{ p: 0, m: 0, textAlign: "center", width: "100%" }}>
+                {children}
+            </Box>
         </Tooltip>
     );
 
 const hiddenStatuses = ["APPROVED", "REJECTED", "BANNED", "SUSPENDED"];
 
-const IndicateNhseSdeAccessButton = ({ sx }: { sx?: SxProps }) => {
+const IndicateNhseSdeAccessButton = ({
+    sx,
+    action,
+    refetchCohort,
+}: {
+    sx?: SxProps;
+    action?: () => void;
+    refetchCohort?: () => void;
+}) => {
     const t = useTranslations("components.IndicateNhseSdeAccessButton");
+    const { showModal } = useModal();
     const { isLoggedIn, user, isLoading: isLoadingAuth } = useAuth();
     const { isNhsSdeApplicationsEnabled } = useFeatures();
     const { nhseSdeRequestStatus, isLoading } = useCohortStatus(user?.id, {
@@ -40,14 +53,34 @@ const IndicateNhseSdeAccessButton = ({ sx }: { sx?: SxProps }) => {
     );
 
     const onClick = async () => {
-        setHasClickedButton(true);
+        if (action) {
+            return action();
+        }
+
+        setHasClickedButton(true); // move to modal action
+
+        showModal({
+            content: (
+                <Stack sx={{ gap: 2, textAlign: "center" }}>
+                    <Typography
+                        variant="h2"
+                        sx={{ mt: 2, mb: 1 }}
+                        color="secondary">
+                        {t("successTitle")}
+                    </Typography>
+                    <Typography>{t("successInfo")}</Typography>
+                </Stack>
+            ),
+            showConfirm: false,
+            showCancel: false,
+        });
+
         const result = await submitRequest({});
         if (result) {
             revalidateCacheAction(`cohort-user-${user?.id}`);
-            notificationService.apiSuccess(t("success"));
+            refetchCohort && refetchCohort();
         }
     };
-
     const approvalPending =
         hasClickedButton || nhseSdeRequestStatus === "APPROVAL REQUESTED";
     const isDisabled =
@@ -70,9 +103,9 @@ const IndicateNhseSdeAccessButton = ({ sx }: { sx?: SxProps }) => {
     if (isNhsSdeApplicationsEnabled) {
         return (
             <ConditionalWrapper requiresWrapper={isDisabled} wrapper={wrapper}>
-                <div>
+                <span style={{ width: "100%" }}>
                     <Button
-                        sx={{ ...sx, width: "100%" }}
+                        sx={{ ...sx }}
                         variant="outlined"
                         color="secondary"
                         disabled={isDisabled}
@@ -83,7 +116,7 @@ const IndicateNhseSdeAccessButton = ({ sx }: { sx?: SxProps }) => {
                             t("label")
                         )}
                     </Button>
-                </div>
+                </span>
             </ConditionalWrapper>
         );
     }