+++ src/utils/gatewayFlagAdapter.ts
@@ -14,13 +14,17 @@ const getFeatures = async (): Promise<Record<string, boolean>> => {
     try {
         const cookieStore = await cookies();
         const jwtToken = cookieStore?.get(config.JWT_COOKIE)?.value;
+        const hasToken = Boolean(jwtToken);
+        //get my feature flags when logged in, otherwise get global features
+        const url = hasToken ? `${apis.features}/me` : `${apis.features}`;
 
-        //get the features for the current jwt user using /me
-        const res = await fetch(`${apis.features}/me`, {
-            headers: {
-                Authorization: `Bearer ${jwtToken}`,
-            },
+        const res = await fetch(url, {
+            headers: hasToken
+                ? { Authorization: `Bearer ${jwtToken}` }
+                : undefined,
+            cache: "no-store",
         });
+
         if (!res.ok) {
             console.error(`Failed to fetch feature flags: ${res.statusText}`);
             return {};